+++ src/features/post-management/PostFormMain.tsx
@@ -1,47 +1,22 @@
 import styled from "styled-components";
 
-import { BodyTextInput, CodeInput, SummaryInput, TitleInput } from "../../components/post";
-import { usePost } from "../../hooks/usePost";
-import { hasValueProperty } from "../../utils/hasValueProperty";
-import { isSupportedPostAction } from "./isSupportedAction";
-import { PostActionType, PostFormHandler } from "./type";
+import PostFormInput from "./PostFormInput";
+import PostPreview from "./PostPreview";
 import { config } from "./config";
+import { usePost } from "../../hooks/usePost";
 
 const PostFormMain = () => {
   const {
     dispatch,
-    state: { payload },
+    state: { payload, status },
   } = usePost();
 
-  // 사용자 입력 이벤트 처리
-  const handleChange: PostFormHandler = (type) => (e) => {
-    const { target } = e;
-
-    if (!isSupportedPostAction(type)) {
-      throw new Error(`${type} cannot be handled by handleChange.`);
-    }
-
-    if (!hasValueProperty<HTMLInputElement | HTMLTextAreaElement>(target)) {
-      throw new ReferenceError("Value attribute is missing.");
-    }
-
-    dispatch({ type, payload: target.value });
-  };
-
-  const handleCodeChange = (type: PostActionType) => (code?: string) => {
-    if (type === PostActionType.SET_CODE && code !== undefined) {
-      dispatch({ type, payload: code });
-    }
-  };
+  const Component =
+    status === "create" ? <PostFormInput data={payload} dispatch={dispatch} /> : <PostPreview />;
 
   return (
     <OuterContainer>
-      <InnerContainer>
-        <TitleInput onChange={handleChange(PostActionType.SET_TITLE)} value={payload.title} />
-        <CodeInput onChange={handleCodeChange(PostActionType.SET_CODE)} value={payload.code} />
-        <BodyTextInput onChange={handleChange(PostActionType.SET_BODY)} value={payload.body} />
-        <SummaryInput onChange={handleChange(PostActionType.SET_SUMMARY)} value={payload.summary} />
-      </InnerContainer>
+      <InnerContainer>{Component}</InnerContainer>
     </OuterContainer>
   );
 };
