+++ src/features/post-management/PostPreview.tsx
@@ -1,0 +1,5 @@
+const PostPreview = () => {
+  return <div>PostPreview</div>;
+};
+
+export default PostPreview;
