+++ src/features/post-management/index.ts
@@ -1,3 +1,3 @@
-export { default as PostFormMain } from "./PostFormMain";
+export { default as PostFormInput } from "./PostFormInput";
 export { default as PostFormHeader } from "./PostFormHeader";
 export { default as PostFormFooter } from "./PostFormFooter";
