+++ src/pages/PostCreatePage.tsx
@@ -1,9 +1,10 @@
 import { Layout } from "antd";
+import styled from "styled-components";
 
-import { PostFormFooter, PostFormHeader, PostFormMain } from "../features/post-management";
+import { PostFormFooter, PostFormHeader } from "../features/post-management";
 import { PostProvider } from "../features/post-management/PostContext";
 import { config } from "../features/post-management/config";
-import styled from "styled-components";
+import PostFormMain from "../features/post-management/PostFormMain";
 
 const { Sider, Content } = Layout;
 
@@ -11,26 +12,26 @@ const PostCreatePage = () => {
   return (
     <Layout>
       <Sider collapsed></Sider>
-      <Layout>
-        <StyledContent>
-          <PostForm />
-        </StyledContent>
-      </Layout>
+      <PostForm />
     </Layout>
   );
 };
 
 const PostForm = () => {
   return (
-    <PostProvider>
-      <PostFormHeader />
-      <PostFormMain />
-      <PostFormFooter />
-    </PostProvider>
+    <Layout>
+      <Container>
+        <PostProvider>
+          <PostFormHeader />
+          <PostFormMain />
+          <PostFormFooter />
+        </PostProvider>
+      </Container>
+    </Layout>
   );
 };
 
-const StyledContent = styled(Content)`
+const Container = styled(Content)`
   display: flex;
   flex-direction: column;
   height: 100vh;
